+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * New lint `ImplicitDiv` `%div`s which are unnecessary due to a class or ID
   specified on the tag
+* New lint `TagName` ensures tag names are lowercase
 
 ## 0.1.0